# doc/swagger.yaml
info:
  contact: {}
  description: This is a sample mailing servivce.
  title: czwrMailing - profile service
  version: "1.0"
paths:
  /debug/pprof_disable:
    post:
      consumes:
      - application/json
      description: disable pprof API on service
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: disable pprof API
      tags:
      - pprofwrapper
  /debug/pprof_enable:
    post:
      consumes:
      - application/json
      description: enable pprof API on service
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: enable pprof API
      tags:
      - pprofwrapper
  /profile/i:
    get:
      consumes:
      - application/json
      description: get auth data
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: show user info
      tags:
      - profile
  /profile/me:
    get:
      consumes:
      - application/json
      description: get auth data
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: valid jwt token and show user info
      tags:
      - profile
  /profile/receivers/{id}:
    delete:
      consumes:
      - application/json
      description: delete receiver
      parameters:
      - description: USR
        in: query
        name: string
        required: true
        type: string
      - description: RECEIVER
        in: query
        name: string
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: delete receiver from receiver list
      tags:
      - profile
  /profile/receivers/{usr}:
    get:
      consumes:
      - application/json
      description: get receivers
      parameters:
      - description: USR
        in: query
        name: string
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: get receivers list
      tags:
      - profile
  /profile/receivers/{usr}/{id}/{receiver}:
    patch:
      consumes:
      - application/json
      description: update receiver
      parameters:
      - description: USR
        in: query
        name: string
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: RECEIVER
        in: query
        name: string
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: update receiver in receiver list
      tags:
      - profile
  /profile/receivers/{usr}/{receiver}:
    post:
      consumes:
      - application/json
      description: create receiver
      parameters:
      - description: USR
        in: query
        name: string
        required: true
        type: string
      - description: RECEIVER
        in: query
        name: string
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: create receiver in receivers list
      tags:
      - profile
  /profile/upload_template:
    delete:
      consumes:
      - application/json
      description: upload data
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: upload template data
      tags:
      - profile
swagger: "2.0"